# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BMLT",
        "description": "BMLT Admin API Documentation",
        "license": {
            "name": "MIT",
            "url": "https://github.com/bmlt-enabled/bmlt-server/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/servers": {
            "get": {
                "tags": [
                    "server"
                ],
                "summary": "Retrieves servers",
                "description": "Retrieve servers.",
                "operationId": "getServers",
                "responses": {
                    "200": {
                        "description": "Successful response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServerCollection"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Returns when aggregator mode is disabled.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/servers/{serverId}": {
            "get": {
                "tags": [
                    "server"
                ],
                "summary": "Retrieves a server",
                "description": "Retrieve a single server id.",
                "operationId": "getServer",
                "parameters": [
                    {
                        "name": "serverId",
                        "in": "path",
                        "description": "ID of server",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Server"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Returns when no server exists.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NotFoundError": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The requested resource was not found."
                    }
                },
                "type": "object"
            },
...
            "ServerBase": {
                "properties": {
                    "sourceId": {
                        "type": "integer",
                        "example": "0"
                    },
                    "name": {
                        "type": "string",
                        "example": "string"
                    },
                    "url": {
                        "type": "string",
                        "example": "https://example.com/main_server"
                    },
                    "statistics": {
                        "required": [
                            "serviceBodies",
                            "meetings"
                        ],
                        "properties": {
                            "serviceBodies": {
                                "required": [
                                    "numZones",
                                    "numRegions",
                                    "numAreas",
                                    "numGroups"
                                ],
                                "properties": {
                                    "numZones": {
                                        "type": "integer",
                                        "example": "0"
                                    },
                                    "numRegions": {
                                        "type": "integer",
                                        "example": "0"
                                    },
                                    "numAreas": {
                                        "type": "integer",
                                        "example": "0"
                                    },
                                    "numGroups": {
                                        "type": "integer",
                                        "example": "0"
                                    }
                                },
                                "type": "object"
                            },
                            "meetings": {
                                "required": [
                                    "numTotal",
                                    "numInPerson",
                                    "numVirtual",
                                    "numHybrid",
                                    "numUnknown"
                                ],
                                "properties": {
                                    "numTotal": {
                                        "type": "integer",
                                        "example": "0"
                                    },
                                    "numInPerson": {
                                        "type": "integer",
                                        "example": "0"
                                    },
                                    "numVirtual": {
                                        "type": "integer",
                                        "example": "0"
                                    },
                                    "numHybrid": {
                                        "type": "integer",
                                        "example": "0"
                                    },
                                    "numUnknown": {
                                        "type": "integer",
                                        "example": "0"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "serverInfo": {
                        "type": "string",
                        "example": "string"
                    },
                    "lastSuccessfulImport": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-11-25 04:16:26"
                    }
                },
                "type": "object"
            },
            "Server": {
                "required": [
                    "id",
                    "sourceId",
                    "name",
                    "url",
                    "lastSuccessfulImport"
                ],
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ServerBase"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": "0"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "ServerCollection": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/Server"
                }
            },
...
        },
        "securitySchemes": {
            "bmltToken": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "api/v1/auth/token",
                        "refreshUrl": "api/v1/auth/refresh",
                        "scopes": {}
                    }
                }
            }
        }
    },
...
}